Update Scripts
Remove module loading and unloading in start and Revert scripts. Pointed out by McGyverNL it's probably not required, and I can't honestly remember why I had it in there in the first place. I imagine that there was actually a reason, but I need to take better notes. That was a year and a half ago.